Move SlotKey type to core
